% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMCensSkewReg-package.R
\docType{package}
\name{FMCensSkewReg-package}
\alias{FMCensSkewReg}
\alias{FMCensSkewReg-package}
\title{FMCensSkewReg: Finite Mixture of Censored Regression with Skewed Distributions}
\description{
Functions to fit finite-mixture censored regression models under
Normal ("Normal"), Student-t ("T"), Skew-Normal ("SN"), and Skew-t ("ST").
Supports left-censoring, k-means initialization, Aitken acceleration,
and optional information-matrix standard errors.
}
\section{Main function}{

\itemize{
\item \code{\link{EM.skewCens.mixR}}: EM algorithm for finite-mixture censored regression.
}
}

\author{
\strong{Maintainer}: Jiwon Park \email{pcjylove87@gmail.com}

Authors:
\itemize{
  \item Victor Hugo Lachos Davila
}

}
\keyword{internal}
